+++ components/level/sidebar.tsx
@@ -1,6 +1,7 @@
 import React, { useContext } from 'react';
 import Dimensions from '../../constants/dimensions';
 import FormattedLevelInfo from '../formattedLevelInfo';
+import FormattedLevelReviews from '../formattedLevelReviews';
 import { LevelContext } from '../../contexts/levelContext';
 import { PageContext } from '../../contexts/pageContext';
 import cleanAuthorNote from '../../helpers/cleanAuthorNote';
@@ -8,12 +9,6 @@ import cleanAuthorNote from '../../helpers/cleanAuthorNote';
 export default function Sidebar() {
   const levelContext = useContext(LevelContext);
   const { windowSize } = useContext(PageContext);
-
-  if (!levelContext) {
-    return null;
-  }
-
-  const { level, reviews } = levelContext;
 
   return (
     <div className='border-l p-4' style={{
@@ -24,14 +19,21 @@ export default function Sidebar() {
       right: 0,
       width: Dimensions.SidebarWidth,
     }}>
-      {!level?.authorNote ? null :
-        <div className='mb-4'>
-          <span style={{ whiteSpace: 'pre-wrap' }}>
-            {cleanAuthorNote(level.authorNote)}
-          </span>
-        </div>
+      {!levelContext?.level ? null :
+        <>
+          {!levelContext.level.authorNote ? null :
+            <div className='mb-4'>
+              <span style={{ whiteSpace: 'pre-wrap' }}>
+                {cleanAuthorNote(levelContext.level.authorNote)}
+              </span>
+            </div>
+          }
+          <div className='mb-4'>
+            <FormattedLevelInfo level={levelContext.level} />
+          </div>
+          <FormattedLevelReviews levelId={levelContext.level._id.toString()} />
+        </>
       }
-      {!level ? null : <FormattedLevelInfo level={level} />}
     </div>
   );
 }
